fix: broken dapp favicon in connected websites list
